fix(web): satisfy homepage payload parser CI budget
Summary:
- extract workspace payload parsing primitives into a dedicated helper
  module so the homepage normalization logic stays under the web file
  budget gate
- rename the HR advisor fresh-project regression test to match the
  default-recommendation behavior it asserts

Rationale:
- the PR was mergeable but blocked by the frontend CI hard limit on
  web/src/lib/features/workspace/payloads.ts after the normalization
  work landed
- aligning the test name also resolves the only actionable review
  comment on the PR without changing runtime behavior

Tests:
- corepack pnpm --dir web run ci
- PATH=/home/yuzhong/.local/go1.26.1/bin:$PATH go test ./internal/httpapi -run 'Test(ListAgentsRouteReturnsEmptyCapabilitiesArray|ListActivityEventsRouteReturnsEmptyArrayWhenNoEventsExist|HRAdvisorRouteReturnsDefaultRecommendationsForFreshProject|ListTicketStatusesRouteReturnsEmptyArrayForNewProject|ListWorkflowsRouteReturnsEmptyArrayForNewProject|ListTicketsRouteReturnsEmptyArrayForNewProject)$'
- ./scripts/ci/lint.sh ./internal/httpapi/...
- PATH=/home/yuzhong/.local/go1.26.1/bin:$PATH go build -o /tmp/openase-issue-92 ./cmd/openase
